Fixed DB issues due to mismatching schema in prepackaged DB and incorrect SQL queries in DAOs.
