+++ app/src/main/java/edu/card/clarity/data/creditCard/CreditCardDao.kt
@@ -4,6 +4,9 @@ import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Upsert
+import edu.card.clarity.data.creditCard.predefined.PredefinedCreditCard
+import edu.card.clarity.data.creditCard.userAdded.UserAddedCreditCard
+import edu.card.clarity.data.creditCard.userAdded.UserAddedCreditCardInfo
 import edu.card.clarity.enums.RewardType
 import kotlinx.coroutines.flow.Flow
 import java.util.UUID
@@ -15,16 +18,16 @@ interface CreditCardDao {
      *
      * @return all credit cards.
      */
-    @Query("SELECT * FROM creditCardInfo")
-    fun observeAllInfo(): Flow<List<CreditCardInfoEntity>>
+    @Query("SELECT * FROM userAddedCreditCard")
+    fun observeAllInfo(): Flow<List<UserAddedCreditCardInfo>>
 
     /**
      * Observe list of all credit cards of a certain reward type.
      *
      * @return all credit cards of a certain reward type.
      */
-    @Query("SELECT * FROM creditCardInfo WHERE rewardType = :type")
-    fun observeAllInfoOf(type: RewardType): Flow<List<CreditCardInfoEntity>>
+    @Query("SELECT * FROM userAddedCreditCard WHERE rewardType = :rewardType")
+    fun observeAllInfoOf(rewardType: RewardType): Flow<List<UserAddedCreditCardInfo>>
 
     /**
      * Observe a single credit card's info and its associated benefits by id.
@@ -43,7 +46,7 @@ interface CreditCardDao {
      * @return the credit card with id.
      */
     @Query("SELECT * FROM creditCardInfo WHERE id = :id")
-    fun observeInfoById(id: UUID): Flow<CreditCardInfoEntity>
+    fun observeInfoById(id: UUID): Flow<CreditCardInfo>
 
     /**
      * Observe a single credit card's reward type by id.
@@ -59,8 +62,9 @@ interface CreditCardDao {
      *
      * @return all credit cards with credit cards associated to them.
      */
-    @Query("SELECT * FROM creditCardInfo")
-    fun observeAll(): Flow<List<CreditCardInfoEntity>>
+    @Transaction
+    @Query("SELECT * FROM userAddedCreditCard")
+    fun observeAll(): Flow<List<UserAddedCreditCard>>
 
     /**
      * Observe all credit cards of a certain reward type.
@@ -68,8 +72,9 @@ interface CreditCardDao {
      * @return all credit cards of a certain reward type with credit cards associated to them.
      */
     @Transaction
-    @Query("SELECT * FROM creditCardInfo WHERE rewardType = :rewardType")
-    fun observeAllOf(rewardType: RewardType): Flow<List<CreditCardEntity>>
+    @Query("SELECT * FROM userAddedCreditCard WHERE rewardType = :rewardType"
+    )
+    fun observeAllOf(rewardType: RewardType): Flow<List<UserAddedCreditCard>>
 
     /**
      * Observe all predefined credit cards of a certain reward type.
@@ -77,28 +82,24 @@ interface CreditCardDao {
      * @return all such predefined credit cards with credit cards associated to them .
      */
     @Transaction
-    @Query("SELECT creditCardInfo.* FROM creditCardInfo " +
-            "JOIN predefinedCreditCardId " +
-            "ON creditCardInfo.id = predefinedCreditCardId.creditCardId " +
-            "WHERE creditCardInfo.rewardType = :rewardType"
-    )
-    fun observeAllPredefinedOf(rewardType: RewardType): Flow<List<CreditCardEntity>>
+    @Query("SELECT * FROM predefinedCreditCard WHERE rewardType = :rewardType")
+    fun observeAllPredefinedOf(rewardType: RewardType): Flow<List<PredefinedCreditCard>>
 
     /**
      * Select all credit cards from the credit card table.
      *
      * @return all credit cards.
      */
-    @Query("SELECT * FROM creditCardInfo")
-    suspend fun getAllInfo(): List<CreditCardInfoEntity>
+    @Query("SELECT * FROM userAddedCreditCard")
+    suspend fun getAllInfo(): List<UserAddedCreditCardInfo>
 
     /**
      * Select all credit cards of a certain reward type.
      *
      * @return all credit cards.
      */
-    @Query("SELECT * FROM creditCardInfo WHERE rewardType = :rewardType")
-    suspend fun getAllInfoOf(rewardType: RewardType): List<CreditCardInfoEntity>
+    @Query("SELECT * FROM userAddedCreditCard WHERE rewardType = :rewardType")
+    suspend fun getAllInfoOf(rewardType: RewardType): List<UserAddedCreditCardInfo>
 
     /**
      * Select a credit card by id.
@@ -107,7 +108,7 @@ interface CreditCardDao {
      * @return the credit card with id.
      */
     @Query("SELECT * FROM creditCardInfo WHERE id = :id")
-    suspend fun getInfoById(id: UUID): CreditCardInfoEntity?
+    suspend fun getInfoById(id: UUID): CreditCardInfo?
 
     /**
      * Select a single credit card's reward type by id.
@@ -124,8 +125,8 @@ interface CreditCardDao {
      * @return all such credit cards with associated rewards.
      */
     @Transaction
-    @Query("SELECT * FROM creditCardInfo WHERE rewardType = :rewardType")
-    suspend fun getAllOf(rewardType: RewardType): List<CreditCardEntity>
+    @Query("SELECT * FROM userAddedCreditCard WHERE rewardType = :rewardType")
+    suspend fun getAllOf(rewardType: RewardType): List<UserAddedCreditCard>
 
     /**
      * Select all predefined credit cards of a certain reward type.
@@ -133,12 +134,8 @@ interface CreditCardDao {
      * @return all such predefined credit cards with credit cards associated to them.
      */
     @Transaction
-    @Query("SELECT creditCardInfo.* FROM creditCardInfo " +
-            "JOIN predefinedCreditCardId " +
-            "ON creditCardInfo.id = predefinedCreditCardId.creditCardId " +
-            "WHERE creditCardInfo.rewardType = :rewardType"
-    )
-    suspend fun getAllPredefinedOf(rewardType: RewardType): List<CreditCardEntity>
+    @Query("SELECT * FROM predefinedCreditCard WHERE rewardType = :rewardType")
+    suspend fun getAllPredefinedOf(rewardType: RewardType): List<PredefinedCreditCard>
 
     /**
      * Select a credit card with associated purchase rewards by id.
@@ -148,7 +145,7 @@ interface CreditCardDao {
      */
     @Transaction
     @Query("SELECT * FROM creditCardInfo WHERE id = :id")
-    suspend fun getById(id: UUID): CreditCardEntity?
+    suspend fun getById(id: UUID): UserAddedCreditCard?
 
     /**
      * Insert or update a credit card in the database. If a credit card already exists, replace it.
@@ -156,7 +153,7 @@ interface CreditCardDao {
      * @param cashBackCreditCard the credit card to be inserted or updated.
      */
     @Upsert
-    suspend fun upsert(cashBackCreditCard: CreditCardInfoEntity)
+    suspend fun upsert(cashBackCreditCard: CreditCardInfo)
 
     /**
      * Insert or update credit cards in the database. If a credit card already exists, replace it.
@@ -164,7 +161,7 @@ interface CreditCardDao {
      * @param cashBackCreditCards the credit cards to be inserted or updated.
      */
     @Upsert
-    suspend fun upsertAll(cashBackCreditCards: List<CreditCardInfoEntity>)
+    suspend fun upsertAll(cashBackCreditCards: List<CreditCardInfo>)
 
     /**
      * Delete a credit card by id.
