+++ app/src/main/java/edu/card/clarity/data/creditCard/ICreditCard.kt
@@ -1,0 +1,12 @@
+package edu.card.clarity.data.creditCard
+
+import edu.card.clarity.data.purchaseReward.PurchaseRewardEntity
+
+/**
+ * A database interface that mirrors [edu.card.clarity.domain.creditCard.ICreditCard],
+ * which contains both credit card information and associated purchase rewards.
+ */
+interface ICreditCard {
+    val creditCardInfo: ICreditCardInfo
+    val purchaseRewards: List<PurchaseRewardEntity>
+}
