+++ app/src/main/java/edu/card/clarity/data/creditCard/userAdded/UserAddedCreditCard.kt
@@ -1,0 +1,15 @@
+package edu.card.clarity.data.creditCard.userAdded
+
+import androidx.room.Embedded
+import androidx.room.Relation
+import edu.card.clarity.data.creditCard.ICreditCard
+import edu.card.clarity.data.purchaseReward.PurchaseRewardEntity
+
+data class UserAddedCreditCard(
+    @Embedded override val creditCardInfo: UserAddedCreditCardInfo,
+    @Relation(
+        parentColumn = "id",
+        entityColumn = "creditCardId"
+    )
+    override val purchaseRewards: List<PurchaseRewardEntity>
+): ICreditCard
