+++ app/src/main/java/edu/card/clarity/data/purchaseReward/PurchaseRewardEntity.kt
@@ -3,7 +3,7 @@ package edu.card.clarity.data.purchaseReward
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
-import edu.card.clarity.data.creditCard.CreditCardInfoEntity
+import edu.card.clarity.data.creditCard.CreditCardInfo
 import edu.card.clarity.enums.PurchaseType
 import edu.card.clarity.enums.RewardType
 import java.util.UUID
@@ -13,7 +13,7 @@ import java.util.UUID
     primaryKeys = ["creditCardId", "purchaseType"],
     foreignKeys = [
         ForeignKey(
-            entity = CreditCardInfoEntity::class,
+            entity = CreditCardInfo::class,
             parentColumns = ["id"],
             childColumns = ["creditCardId"],
             onDelete = ForeignKey.CASCADE
