+++ app/src/test/java/edu/card/clarity/domain/creditCard/PointBackUserAddedCreditCardTest.kt
@@ -17,7 +17,7 @@ import org.mockito.Mockito.mockStatic
 import java.util.Date
 import java.util.UUID
 
-class PointBackCreditCardTest {
+class PointBackUserAddedCreditCardTest {
 
     private val mockCalendar = mock(Calendar::class.java)
     private val mockedStaticCalendar: MockedStatic<Calendar> = mockStatic(Calendar::class.java)
